优化ES索引数据类型

+++ src/main/java/net/diaowen/dwsurvey/entity/es/answer/question/EsAnQuestion.java
@@ -13,7 +13,7 @@ public class EsAnQuestion {
     //题型，使用QuType类名称来标识(全大写)
     private String quType;
     //题目答卷JSON，注意针对不同题型答卷JSON结构可能不一样。
-    private Double quAnScore; // 得分
+    private float quAnScore; // 得分
     //单选题
     private EsAnRadio anRadio;
     //多选题
@@ -111,11 +111,11 @@ public class EsAnQuestion {
         this.anOtherMaps = anOtherMaps;
     }
 
-    public Double getQuAnScore() {
+    public float getQuAnScore() {
         return quAnScore;
     }
 
-    public void setQuAnScore(Double quAnScore) {
+    public void setQuAnScore(float quAnScore) {
         this.quAnScore = quAnScore;
     }
 }
